// apps/api/src/app/shared/enum/error-type.enum.ts
export enum ErrorType {
    HTTP_EXCEPTION = 'HTTP_EXCEPTION',
    UNIQUE_CONSTRAINT_VIOLATION = 'UNIQUE_CONSTRAINT_VIOLATION',
    DATABASE_ERROR = 'DATABASE_ERROR',
    ENTITY_NOT_FOUND = 'ENTITY_NOT_FOUND',
    INTERNAL_SERVER_ERROR = 'INTERNAL_SERVER_ERROR',
    FORBIDDEN = "FORBIDDEN",
    UNAUTHORIZED = "UNAUTHORIZED"
}